changed all np.float64, double, to np.float32, float, respectively

+++ peripy/correction.pxd
@@ -1,3 +1,3 @@
-cdef double cPA_HHB(double[:], double[:], double, double)
+cdef float cPA_HHB(float[:], float[:], float, float)
 
-cdef double cmicromodulus_connical(double[:], double[:], double)
+cdef float cmicromodulus_connical(float[:], float[:], float)
