+++ peripy/test/test_cl_euler_cromer.py
@@ -37,14 +37,14 @@ class TestUpdateDisplacement:
     @context_available
     def test_update_displacement(self, context, queue, program):
         """Test basic displacement update."""
-        u = np.zeros(3, dtype=np.float64)
-        ud = np.zeros(3, dtype=np.float64)
-        udd = np.zeros(3, dtype=np.float64)
-        densities = np.ones(3, dtype=np.float64)
+        u = np.zeros(3, dtype=np.float32)
+        ud = np.zeros(3, dtype=np.float32)
+        udd = np.zeros(3, dtype=np.float32)
+        densities = np.ones(3, dtype=np.float32)
         nnodes = 1
-        force = np.array([1.0, 2.0, 3.0], dtype=np.float64)
+        force = np.array([1.0, 2.0, 3.0], dtype=np.float32)
         bc_types = np.array([0, 0, 0], dtype=np.intc)
-        bc_values = np.array([0, 0, 0], dtype=np.float64)
+        bc_values = np.array([0, 0, 0], dtype=np.float32)
         displacement_bc_scale = 0
         dt = 1.0
         damping = 1.0
@@ -80,8 +80,8 @@ class TestUpdateDisplacement:
         update_displacement_kernel(
             queue, (3 * nnodes,), None,
             force_d, u_d, ud_d, udd_d, bc_types_d, bc_values_d, densities_d,
-            np.float64(displacement_bc_scale), np.float64(damping),
-            np.float64(dt)
+            np.float32(displacement_bc_scale), np.float32(damping),
+            np.float32(dt)
             )
         cl.enqueue_copy(queue, u, u_d)
         cl.enqueue_copy(queue, ud, ud_d)
@@ -94,14 +94,14 @@ class TestUpdateDisplacement:
     @context_available
     def test_update_displacement2(self, context, queue, program):
         """Test displacement update."""
-        u = np.zeros(3, dtype=np.float64)
-        ud = np.zeros(3, dtype=np.float64)
-        udd = np.zeros(3, dtype=np.float64)
-        densities = np.ones(3, dtype=np.float64)
+        u = np.zeros(3, dtype=np.float32)
+        ud = np.zeros(3, dtype=np.float32)
+        udd = np.zeros(3, dtype=np.float32)
+        densities = np.ones(3, dtype=np.float32)
         nnodes = 1
-        force = np.array([1.0, 2.0, 3.0], dtype=np.float64)
+        force = np.array([1.0, 2.0, 3.0], dtype=np.float32)
         bc_types = np.array([0, 0, 0], dtype=np.intc)
-        bc_values = np.array([0, 0, 0], dtype=np.float64)
+        bc_values = np.array([0, 0, 0], dtype=np.float32)
         displacement_bc_scale = 0
         dt = 2.0
         damping = 1.0
@@ -137,8 +137,8 @@ class TestUpdateDisplacement:
         update_displacement_kernel(
             queue, (3 * nnodes,), None,
             force_d, u_d, ud_d, udd_d, bc_types_d, bc_values_d, densities_d,
-            np.float64(displacement_bc_scale), np.float64(damping),
-            np.float64(dt)
+            np.float32(displacement_bc_scale), np.float32(damping),
+            np.float32(dt)
             )
         cl.enqueue_copy(queue, u, u_d)
         cl.enqueue_copy(queue, ud, ud_d)
@@ -151,14 +151,14 @@ class TestUpdateDisplacement:
     @context_available
     def test_update_displacement3(self, context, queue, program):
         """Test displacement update with displacement boundary conditions."""
-        u = np.zeros(3, dtype=np.float64)
-        ud = np.array([1.0, 1.0, 1.0], dtype=np.float64)
-        udd = np.zeros(3, dtype=np.float64)
-        densities = np.array([1.0, 1.0, 1.0], dtype=np.float64)
+        u = np.zeros(3, dtype=np.float32)
+        ud = np.array([1.0, 1.0, 1.0], dtype=np.float32)
+        udd = np.zeros(3, dtype=np.float32)
+        densities = np.array([1.0, 1.0, 1.0], dtype=np.float32)
         nnodes = 1
-        force = np.array([1.0, 2.0, 3.0], dtype=np.float64)
+        force = np.array([1.0, 2.0, 3.0], dtype=np.float32)
         bc_types = np.array([1, 1, 0], dtype=np.intc)
-        bc_values = np.array([0.0, 0.0, 0.0], dtype=np.float64)
+        bc_values = np.array([0.0, 0.0, 0.0], dtype=np.float32)
         displacement_bc_scale = 1.0
         dt = 2.0
         damping = 2.0
@@ -194,15 +194,15 @@ class TestUpdateDisplacement:
         update_displacement_kernel(
             queue, (3 * nnodes,), None,
             force_d, u_d, ud_d, udd_d, bc_types_d, bc_values_d, densities_d,
-            np.float64(displacement_bc_scale), np.float64(damping),
-            np.float64(dt)
+            np.float32(displacement_bc_scale), np.float32(damping),
+            np.float32(dt)
             )
         cl.enqueue_copy(queue, u, u_d)
         cl.enqueue_copy(queue, ud, ud_d)
         cl.enqueue_copy(queue, udd, udd_d)
-        u_expected = np.array([0.0, 0.0, 6.0], dtype=np.float64)
-        ud_expected = np.array([-1.0, 1.0, 3.0], dtype=np.float64)
-        udd_expected = np.array([-1.0, 0.0, 1.0], dtype=np.float64)
+        u_expected = np.array([0.0, 0.0, 6.0], dtype=np.float32)
+        ud_expected = np.array([-1.0, 1.0, 3.0], dtype=np.float32)
+        udd_expected = np.array([-1.0, 0.0, 1.0], dtype=np.float32)
 
         assert np.all(u == u_expected)
         assert np.all(ud == ud_expected)
@@ -211,14 +211,14 @@ class TestUpdateDisplacement:
     @context_available
     def test_update_displacement4(self, context, queue, program):
         """Test displacement update with displacement B.C. scale."""
-        u = np.zeros(3, dtype=np.float64)
-        ud = np.zeros(3, dtype=np.float64)
-        udd = np.zeros(3, dtype=np.float64)
-        densties = np.ones(3, dtype=np.float64)
+        u = np.zeros(3, dtype=np.float32)
+        ud = np.zeros(3, dtype=np.float32)
+        udd = np.zeros(3, dtype=np.float32)
+        densties = np.ones(3, dtype=np.float32)
         nnodes = 1
-        force = np.array([1.0, 2.0, 3.0], dtype=np.float64)
+        force = np.array([1.0, 2.0, 3.0], dtype=np.float32)
         bc_types = np.array([1, 1, 0], dtype=np.intc)
-        bc_values = np.array([2.0, 2.0, 0.0], dtype=np.float64)
+        bc_values = np.array([2.0, 2.0, 0.0], dtype=np.float32)
         displacement_bc_scale = 0.5
         dt = 2.0
         damping = 1.0
@@ -254,14 +254,14 @@ class TestUpdateDisplacement:
         update_displacement_kernel(
             queue, (3 * nnodes,), None,
             force_d, u_d, ud_d, udd_d, bc_types_d, bc_values_d, densities_d,
-            np.float64(displacement_bc_scale), np.float64(damping),
-            np.float64(dt)
+            np.float32(displacement_bc_scale), np.float32(damping),
+            np.float32(dt)
             )
         cl.enqueue_copy(queue, u, u_d)
         cl.enqueue_copy(queue, ud, ud_d)
         cl.enqueue_copy(queue, udd, udd_d)
-        u_expected = np.array([1.0, 1.0, 12.0], dtype=np.float64)
-        ud_expected = np.array([2.0, 4.0, 6.0], dtype=np.float64)
+        u_expected = np.array([1.0, 1.0, 12.0], dtype=np.float32)
+        ud_expected = np.array([2.0, 4.0, 6.0], dtype=np.float32)
         udd_expected = force
 
         assert np.all(u == u_expected)
